remove `new` from `yarn create comit-app new foo`
