+++ new_project/examples/btc_eth/package.json
@@ -7,7 +7,7 @@
         "check": "tsc && tslint -p . && prettier --check '**/*.{ts,json,yml}'",
         "swap": "ts-node ./src/index.ts",
         "fix": "tslint -p . --fix && prettier --write '**/*.{ts,js,json,yml}'",
-        "start-env": "create-comit-app start-env"
+        "start-env": "start-env"
     },
     "keywords": [],
     "author": "",
