+++ new_project/examples/separate_apps/package.json
@@ -8,7 +8,7 @@
         "maker": "ts-node ./src/maker.ts",
         "taker": "ts-node ./src/taker.ts",
         "fix": "tslint -p . --fix && prettier --write '**/*.{ts,js,json,yml}'",
-        "start-env": "create-comit-app start-env"
+        "start-env": "start-env"
     },
     "keywords": [],
     "author": "",
