+++ new_project/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node ./src/index.js",
-    "start-env": "create-comit-app start-env"
+    "start-env": "start-env"
   },
   "dependencies": {
     "comit-sdk": "^0.7.1",
